Update Utils.cs
* Rename to NSprites (file) + NSpritesUtils (class). Can't be name both NSprites, because source generation breakes when namespace and class name are the same
* Implement helper extension methods to safely add all necessary render components to entity during baking (convertion) process or runtime
* Add summary description to already existed methods
